Adding support for mounted folders
When an application on the white list was stored in a mounted folder (NTFS), the client application picked the volume with the mount point while the driver received notifications that the application resided on the mounted volume. This change set ensures that both client and driver refer to the same (mounted) volume.