Remove I/O queue
The filter driver implemented a parallel I/O queue for dispatching device control messages to a handler that was there for tracing purposes only.  Some users were concerned about the potential performance impact. In the current development stage, the added value for tracing is minimal. Removing the I/O queue allows the O/S to bypass the filter driver. This voids the earlier mentioned performance concern.