UI Improvements
This change set incorporates UI improvements related to the Whitelist dialog. Upon startup of the configuration client an attempt will be made to add the configuration client itself to the whitelist. As such the client application is always granted access to hidden HID devices when the Devices tab is opened. In addition, the Whitelist dialog will now support drag and drop of applications on the Whitelist. This feature is limited to .exe and .com and .bin files that reside on a volume.