Initial implementation of CLI
This change set attempts to ease the integration of HidHide with third-party software. As this landscape is diverse and includes other programming languages and scripting languages, a common ground was found for a Command Line Interface (CLI). This change set provides an initial implementation of the CLI and modifies the Configuration Client to handle the device-in-use scenario when the CLI is kept open while the GUI client is started. Over time it is the intention to use the CLI also from within the GUI client, and hence remove the current level of code duplication. This is not done in the current change set, to allow the chance for testing the CLI while keeping the alternative stable.