Duplicate code removal
This change set addresses the code duplication in the GUI introduced with the initial implementation of the CLI. In addition, the GUI will now, just like the CLI, maintain its exclusive lock on the filter driver. Both GUI and CLI are now able to detect registered applications that are no longer present at the registered location.